gzip: also honor GZIP=--rsyncable
* gzip.c (main): Also accept --rsyncable when it is specified
via the GZIP environment variable.
I noticed this when gzip's "make dist" failed because maint.mk
detected that gzip now honors --rsyncable, yet when it set GZIP_ENV
to include that, and that propagated via automake-generated code
to the GZIP setting used in the "make dist" rule, there, it was not
honored, and caused "make dist" to fail.