create rust CI (#2)
This is a quick PR to add a Rust Action to build the project. For now, I've commented out the test portion (since my understanding is that there are no tests).

Maybe eventually we can add an action to benchmark ourselves against the official test suite, and track our progress?